+++ movies/src/main/java/ru/gaket/themoviedb/data/common/KeyboardManager.kt
@@ -1,0 +1,5 @@
+package ru.gaket.themoviedb.data.common
+
+interface KeyboardManager {
+  fun hideKeyboard()
+}
