+++ movies/src/main/java/ru/gaket/themoviedb/data/common/Navigator.kt
@@ -1,0 +1,6 @@
+package ru.gaket.themoviedb.data.common
+
+interface Navigator {
+  fun forward(screen: Screen)
+  fun back()
+}
