+++ movies/src/main/java/ru/gaket/themoviedb/data/common/Screen.kt
@@ -1,0 +1,6 @@
+package ru.gaket.themoviedb.data.common
+
+sealed class Screen {
+    object Movies : Screen()
+    class MovieDetails(val movieId: Int) : Screen()
+}
