+++ movies/src/main/java/ru/gaket/themoviedb/data/entities/Movie.kt
@@ -1,0 +1,6 @@
+package ru.gaket.themoviedb.data.entities
+
+/**
+ * Business class of Movies
+ */
+data class Movie(val id: Int, val name: String, val thumbnail: String?)
