+++ movies/src/main/java/ru/gaket/themoviedb/data/network/SearchMovieResponse.kt
@@ -1,0 +1,10 @@
+package ru.gaket.themoviedb.data.network
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchMovieResponse(
+    @SerializedName("page")
+    val page: Int,
+    @SerializedName("results")
+    val movies: List<MovieNetworkModel>
+)
