--- movies/src/main/java/ru/gaket/themoviedb/model/common/KeyboardManager.kt
@@ -1,5 +1,0 @@
-package ru.gaket.themoviedb.model.common
-
-interface KeyboardManager {
-  fun hideKeyboard()
-}
