--- movies/src/main/java/ru/gaket/themoviedb/model/common/Navigator.kt
@@ -1,6 +1,0 @@
-package ru.gaket.themoviedb.model.common
-
-interface Navigator {
-  fun forward(screen: Screen)
-  fun back()
-}
