--- movies/src/main/java/ru/gaket/themoviedb/model/common/Screen.kt
@@ -1,6 +1,0 @@
-package ru.gaket.themoviedb.model.common
-
-sealed class Screen {
-    object Movies : Screen()
-    class MovieDetails(val movieId: Int) : Screen()
-}
