--- movies/src/main/java/ru/gaket/themoviedb/model/entities/Movie.kt
@@ -1,6 +1,0 @@
-package ru.gaket.themoviedb.model.entities
-
-/**
- * Business class of Movies
- */
-data class Movie(val id: Int, val name: String, val thumbnail: String?)
