--- movies/src/main/java/ru/gaket/themoviedb/model/network/SearchMovieResponse.kt
@@ -1,10 +1,0 @@
-package ru.gaket.themoviedb.model.network
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchMovieResponse(
-    @SerializedName("page")
-    val page: Int,
-    @SerializedName("results")
-    val movies: List<MovieNetworkModel>
-)
