+++ movies/src/main/java/ru/gaket/themoviedb/screens/movies/mvu/MoviesFragment.kt
@@ -17,6 +17,7 @@ import ru.gaket.themoviedb.screens.movies.common.GridSpacingItemDecoration
 import ru.gaket.themoviedb.screens.movies.common.MoviesAdapter
 import ru.gaket.themoviedb.utils.afterTextChanged
 import ru.gaket.themoviedb.utils.hideKeyboard
+import java.time.LocalTime
 
 
 class MoviesFragment : TeaFragment<MoviesFeature.State, MoviesFeature.Message, MoviesFeature.Dependencies>() {
@@ -27,56 +28,29 @@ class MoviesFragment : TeaFragment<MoviesFeature.State, MoviesFeature.Message, M
 
   private var _binding: MoviesFragmentBinding? = null
   private lateinit var moviesAdapter: MoviesAdapter
+  // This property is only valid between onCreateView and onDestroyView.
+  private val binding get() = _binding!!
 
   override val viewModel: MoviesViewModel by viewModels { getVmFactory() }
 
-  override fun onCreateView(
-    inflater: LayoutInflater, container: ViewGroup?,
-    savedInstanceState: Bundle?
-  ): View {
-    _binding = MoviesFragmentBinding.inflate(inflater, container, false)
-    binding.moviesList.apply {
-      val spanCount =
-        // Set span count depending on layout
-        when (resources.configuration.orientation) {
-          Configuration.ORIENTATION_LANDSCAPE -> 4
-          else -> 2
-        }
-      layoutManager = GridLayoutManager(activity, spanCount)
-      moviesAdapter = MoviesAdapter { movie ->
-        dispatch(MoviesFeature.Message.MovieClicked(movie))
-      }
-      adapter = moviesAdapter
-      addItemDecoration(GridSpacingItemDecoration(spanCount, resources.getDimension(R.dimen.itemsDist).toInt(), true))
-      addOnScrollListener(object : RecyclerView.OnScrollListener() {
-        override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
-          super.onScrollStateChanged(recyclerView, newState)
-          if (newState == SCROLL_STATE_DRAGGING) {
-            recyclerView.hideKeyboard()
-          }
-        }
-      })
-    }
+  override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+    initViews(inflater, container)
     return binding.root
   }
-
-  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-    super.onViewCreated(view, savedInstanceState)
-
-    if (savedInstanceState == null) {
-      dispatch(MoviesFeature.Message.SearchUpdated(""))
-    }
-    binding.searchInput.afterTextChanged { query ->
-      dispatch(MoviesFeature.Message.SearchUpdated(query))
-    }
-  }
-
-  // This property is only valid between onCreateView and onDestroyView.
-  private val binding get() = _binding!!
 
   override fun onDestroyView() {
     super.onDestroyView()
     _binding = null
+  }
+
+  override fun initDispatchers() {
+    binding.searchInput.afterTextChanged { query ->
+      dispatch(MoviesFeature.Message.SearchUpdated(query, LocalTime.now()))
+    }
+    moviesAdapter = MoviesAdapter { movie ->
+      dispatch(MoviesFeature.Message.MovieClicked(movie))
+    }
+    binding.moviesList.adapter = moviesAdapter
   }
 
   override fun render(state: MoviesFeature.State) {
@@ -93,5 +67,28 @@ class MoviesFragment : TeaFragment<MoviesFeature.State, MoviesFeature.Message, M
 
   private fun getVmFactory(): MoviesVmFactory =
     (requireActivity().application as MovieApp).appComponent.moviesVmFactory
+
+  private fun initViews(inflater: LayoutInflater, container: ViewGroup?) {
+    _binding = MoviesFragmentBinding.inflate(inflater, container, false)
+    binding.moviesList.apply {
+      val spanCount =
+        // Set span count depending on layout
+        when (resources.configuration.orientation) {
+          Configuration.ORIENTATION_LANDSCAPE -> 4
+          else -> 2
+        }
+      layoutManager = GridLayoutManager(activity, spanCount)
+
+      addItemDecoration(GridSpacingItemDecoration(spanCount, resources.getDimension(R.dimen.itemsDist).toInt(), true))
+      addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
+          super.onScrollStateChanged(recyclerView, newState)
+          if (newState == SCROLL_STATE_DRAGGING) {
+            recyclerView.hideKeyboard()
+          }
+        }
+      })
+    }
+  }
 
 }
